feat(heureka): adds storeprovider test to ensure store slice connections
